Lay out the dashboard
Used a combination of CSS Grid and Flexbox to lay out the different
sections of the page, as well as many other CSS properties and
HTML tools so as to make it look as responsive and visually appealing
as possible.

Regarding the assets, the color pallete choice was a mix of Slate,
Neutral an Blue found on Tailwind, the icons were SVGs from Material
Desing Icons and the sample profile picture presents my own dog, Benny.